Simplify libwapcaplet and remove context objects.
Remove the lwc_context type from the API and ensure that all strings belong to
the one internment context.  This removes a lot of API and simplifies a lot of
function calls, however it does mean that clients of the library *MUST* be
better at reffing and unreffing strings or it'll explode.


svn path=/trunk/libwapcaplet/; revision=10159